tried to fixed zoom in, at least it's now more or less centeres, also implemented the memcpy in ppgDecompress from 3sxtra's fork

+++ src/common/sprites.h
@@ -27,7 +27,7 @@ typedef struct {
 } TextureVertex;
  */
  typedef struct {
-    short u, v;
+    float u, v;
     uint32_t colour;
     float x, y, z;
 } TextureVertex;
